Includes: MOVEONLY: Move CNodeStateStats from main.h to net.h
