+++ src/qt/peertablemodel.h
@@ -5,7 +5,6 @@
 #ifndef BITCOIN_QT_PEERTABLEMODEL_H
 #define BITCOIN_QT_PEERTABLEMODEL_H
 
-#include "main.h"
 #include "net.h"
 
 #include <QAbstractTableModel>
